APB-10575 organise code before removing redundant classes
