+++ app/uk/gov/hmrc/agentregistrationfrontend/controllers/apply/applicantcontactdetails/EmailAddressController.scala
@@ -103,8 +103,9 @@ extends FrontendController(mcc, actions):
               Some(ApplicantEmailAddress(
                 emailAddress = emailAddress,
                 // avoid unsetting verified status of any unchanged email if we are not ignoring verification
-                isVerified = appConfig.ignoreEmailVerification ||
-                  (emailAddress === details.emailAddress && details.isVerified)
+                isVerified =
+                  appConfig.ignoreEmailVerification ||
+                    (emailAddress === details.emailAddress && details.isVerified)
               ))
             case None =>
               Some(ApplicantEmailAddress(
